Converted teamsData to external JSON
